refactor: Update BalancePage layout and improve account info display
- Adjusted spacing and margins in BalancePage.qml for better visual alignment.
- Changed RowLayout to GridLayout for account information, enhancing responsiveness.
- Updated font sizes and implicit heights for dynamic content adjustment.
- Improved button layout for better usability on different screen sizes.
- Fixed ambiguity in AccountModel's predictBalance method by explicitly casting transaction size.

+++ src/models/AccountModel.cpp
@@ -323,7 +323,8 @@ double AccountModel::predictBalance(const QString &cardNumber, const Transaction
     });
 
     // Use the last N transactions for a more relevant trend (e.g., last 10 or all if fewer)
-    int N = std::min(10, transactions.size());
+    // Explicitly cast one of the arguments to int to resolve ambiguity for std::min
+    int N = std::min(10, static_cast<int>(transactions.size()));
     if (N < 2) return getBalance(cardNumber);
 
     QVector<double> balanceChanges;
